add tests

+++ src/main/java/ua/helpdesk/dao/CategoryDao.java
@@ -1,11 +1,11 @@
 package ua.helpdesk.dao;
 
-import ua.helpdesk.entities.Category;
+import ua.helpdesk.entity.Category;
 
 import java.util.List;
 
 public interface CategoryDao {
-    Category findById(Integer id);
+    Category findById(Long id);
 
     Category findByName(String name);
 
@@ -13,11 +13,11 @@ public interface CategoryDao {
 
     void deleteByName(String name);
 
-    void deleteById(Integer id);
+    void deleteById(Long id);
 
     List<Category> findAllData();
 
     List<Category> findCategoryForService(String serviceName);
 
-    List<Category> findCategoryForService(Integer serviceID);
+    List<Category> findCategoryForService(Long serviceID);
 }
