+++ src/main/java/ua/helpdesk/service/CategoryService.java
@@ -1,12 +1,12 @@
 package ua.helpdesk.service;
 
-import ua.helpdesk.entities.Category;
+import ua.helpdesk.entity.Category;
 
 import java.util.List;
 
 public interface CategoryService {
 
-    Category findById(Integer id);
+    Category findById(Long id);
 
     Category findByName(String name);
 
@@ -16,16 +16,16 @@ public interface CategoryService {
 
     void deleteDataByName(String name);
 
-    void deleteDataById(Integer id);
+    void deleteDataById(Long id);
 
     List<Category> findAllData();
 
-    boolean isDataUnique(Integer id, String name);
+    boolean isDataUnique(Long id, String name);
 
-    boolean isDataUnique(Integer id, String name, Integer type_id);
+    boolean isDataUnique(Long id, String name, Long type_id);
 
     List<Category> findCategoryForService(String serviceName);
 
-    List<Category> findCategoryForService(Integer serviceID);
+    List<Category> findCategoryForService(Long serviceID);
 
 }
